Parallel code without Scheduling and with reduction of counting the even numbers in a large array
My Understanding about the code: 
                                                         This program counts even numbers in a large array of 1000000 randomly generated values using parallel computing with OpenMP to improve performance. It first initializes an array and fills it with random numbers between 0 and 999. The function countEvenParallel() then counts the even numbers using multiple threads, utilizing OpenMP's #pragma omp parallel for directive with reduction(+:count) to ensure correct calculations across threads. And the loop scheduling in it is static because I explicitly don’t define the loop scheduling so Uses the default scheduling strategy. OpenMP divides the loop iterations among threads automatically based on its internal load-balancing mechanism. The way iterations are assigned to threads is not explicitly specified. The program measures execution time using omp_get_wtime() before and after the computation. Finally, it prints the total count of even numbers and the time taken. This parallel approach significantly improves efficiency compared to a sequential implementation by distributing workload across multiple CPU cores.

The Average Execution Time are as follows:
Sum = 0.002381 + 0.008417 + 0.003042 + 0.003255 + 0.002809 + 0.003469 + 0.008420 + 0.012920 + 0.012550 + 0.008639 
Average = sum / 10
Average = 0.006590


Speed up:
Speedup (S) is calculated using the formula:
S=T1/T2
where:
•	T1 is the execution time of the first program(Sequntial-Program) (0.007238),
•	T2 is the execution time of the second program(Parallel-Program)(0.006590).
The speedup of the second program compared to the first program is approximately 1.10×.


Reason behind: 
                          The reason behind that the second program use multi-threading. OpenMP distributes computations across multiple CPU cores, reducing execution time. If the workload is evenly distributed among available threads, some tasks execute concurrently, reducing overall execution time.it takes less time than the sequential program so the speed up with the help of OpenMP parallel for is equal to 1.10x.

If we uncomment the following lines in the above code:
printf("%d ", numbers[i]);
printf("\n");
first of all, there is no need of that to uncomment that lines because it prints the random numbers generated by the rand function but if you uncomment that lines The program will print all 1,000,000 random numbers generated in the array. The output will be very large because every number will be printed on the screen. Printing so many numbers will slow down the execution time significantly.  The performance of parallel execution will be affected also. Uncommenting these lines will make the program behave more sequentially for output, even though the computation of counting even numbers is still parallel.