Parallel code with Static Loop scheduling , reduction and with (4-threads)
Here i my ownself give the code only 4 threads with static loop scheduling and reduction to check that if we reduce the threads then what our best code behaviour will be. its performance increases or decreases while decreasing the numbers of threads from system default threads from 8 to 4-threads .

The Average Execution Time are as follows:
Sum = 0.004588 + 0.004655 + 0.004652 + 0.005248 + 0.004556 + 0.005125 + 0.005187 + 0.005095 + 0.005122 + 0.004539
Average = sum / 10 
Average = 0.004877

so after calculating the average execution time we came into result with that after decreasing the number of threads our performance decreases which is logically correct because in parallelism after decreasing the threads the performance decreases.