Parallel(Reduction, Static Scheduling with 4 threads) with chunk size 400
The parallel implementation of counting even numbers in a large array using OpenMP effectively utilizes static scheduling with a chunk size of 400 and reduction to aggregate results across multiple threads. By setting the number of threads to 4, the workload is evenly distributed, ensuring that each thread processes an equal number of elements, which minimizes thread imbalance. The execution time for 10 runs was recorded, and the average execution time was 0.004574 seconds, demonstrating efficient parallel performance.  However, the low average execution time suggests that the parallel approach significantly optimizes the even number counting task compared to a sequential implementation.