Parallel(Reduction, dynamic Scheduling with 4 threads) with chunk size 400
My understanding about that codes: (Here the static code means Q11)The first code uses static scheduling, which splits the work equally among threads before starting, while the second code uses dynamic scheduling, which assigns work to threads as they finish their tasks. The results show that static scheduling (0.004574 sec) was slightly faster than dynamic scheduling (0.005200 sec). This means that when the work is evenly spread out, static scheduling is more efficient because it avoids extra processing overhead. On the other hand, dynamic scheduling is useful when the work is unpredictable, but in this case, it added a little more time due to the way it assigns tasks during execution.